Introduce custom_request to support all verbs
I believe, it's not useful to support every HTTP verb defined
everywhere with a custom method. But it should be possible to use those
verbs, thus we should introduce a method to trigger requests with a
custom verb. Unfortunately, there is already a public request method, so
I had to choose a different name.