Updates README.rdoc with code example
This PR implements PRs #65, #74 and #78.
Some of the PRs contain similar content and other are no longer
based on the current readme. THis is an attempt to clean up the
readme state. In addition some suggestions for further updates
are part of the PR to raise the discussion about:
List of maintainers and the format (email or github username) and
the copyright text we want to have (or not have).