Sets up focussed tests and fixes inconsistency
In order to focus on single tests during development rspec supports
the option `filter_run_when_matching`.
We agreed, that it is fine for us to support this in #235.

Since it is important to not push :focus tags to the sourcetree,
a leftover has been removed `focus: true`.

Finally an inconsistency in the use of `basic_authorize` and
`authorize` has been fixed.
This has additional value, since this PR is the base for PR #235,
which is supposed to add more tests for digest_authorization.
Since we will have different ways to authorize, it is very helpful
to make the method explicit within our tests.