Avoid create multiple large copies of uploaded file data in memory
Add UploadedFile#append_to(buffer), to append to the given buffer
in chunks using readpartial with an outbuf, rewinding the tempfile
before and after (the tempfile that UploadedFile creates should
always be seekable).  Switch Utils.build_file_part to use this
method.

This should result in a general decrease in memory usage for
large files, and the rewinding fixes #261.

This uses the updated spec from #268.

Co-authored-by: Jeremy Evans <code@jeremyevans.net>