Fix: validate StringIO args on UploadedFile.new
`UploadedFile` includes two initialization strategies - it can either be
built by passing as `content` a file-ish object
(`initialize_from_file_path`) or a `StringIO` instance
(`initialize_from_stringio`).

The initializer switches on the presence of the `original_filename`
kwarg to decide which strategy to use. If present, it assumes the user
wants to provide contents via `StringIO`.

However, as `initialize_from_stringio` suggests initialization should
fail with an `ArgumentError` if `original_filename` isn't provided, it
seems that the init strategy selection is currently incorrect.

Fix by actually switching init strategy on the type of the `content`
argument and add missing test to check the `ArgumentError` is really
raised.