Remove development dependency on Sinatra
Fewer dependencies is better, and this way the maintainer does not
have to be familiar with Sinatra to debug the tests, they only need
to know Rack.

Add Rack::Test::FAKE_APP so that you don't need to create a separate
app for each spec. For the digest specs, use a shared app instead of
creating a separate one for each spec.