Add support for the rack main branch (what will be Rack 3)
Delay loading the mock_digest_request file until it is needed.
This depends on digest auth support that is deprecated in the
rack main branch.  It will probably be removed from rack-test
at some point.

If Rack 3 is used, load only the specific parts of rack that
rack-test uses (request, mock, utils), so this doesn't set
up unnecessary autoloads.

Changes required are fairly small.  The Set-Cookie header in
the specs needs to be lower case.  To deal with the removal
of the rewindable body requirement, add a middleware to the
the specs that resets the initial rack.input in env back to
whatever it was originally. Only a single manual rewind call
was needed after that.

Fix cookie time expiration strings in the specs to use the
expected format.

To get this tested in CI, test on the rack main branch on Ruby
3.1.