Deprecate digest authentication support
This is being deprecated in Rack 3, and rack-test relies on the
Rack implementation, so it needs to be deprecated and then
removed from rack-test.

This deprecates direct access to the constant, as well as
provides appropriate warning messages for calling
digest_authorize both directly on the session and through the
delegator.