Use lower case headers
This is compatible with earlier versions of Rack, and lower-case
headers are required in Rack 3.  Accessing the headers for the
last response uses a case insensitive match.

This also changes the headers used in multipart requests, for
consistency, even though that is not required by the rack SPEC.

While here, remove some bogus comments in the specs about
Rack::ContentLength setting content length for requests.