Refactor and update documentation for Rack::Test::Cookie{,Jar}
Switch Cookie#raw to an attr_reader for simplicitly.

Avoid unnecessary hash allocation in many CookieJar methods
by iterating over @cookies directly, or using newly added
private method each_cookie_for, which yields cookie that
match the URI, instead of building a hash and then accessing
that.

Speed up #for by building a string instead of building a hash,
calling values to get an array, calling map on that, and then
going with a string.