Refactor and update documentation for Rack::Test::Utils
Use extend self instead of making all methods module functions.
By the combined use of module_function and private, it appears
the intention was to make build_nested_query and build_multipart
public, and the rest private. module_function actually made all
methods private instance methods and public class methods,
making the private method call in the module pointless.  extend
self seems to better fit the intent, and has the advantage of
not making methods that should be private public class methods.

Split build_multipart into two methods, one public method and
one private method that can be called recursively.  This
allows for avoiding work not needed in the recursive case.

Add a test for the case where the third argument to
build_multipart is provided.  This was added as a deliberate
feature in e1b0d73a2e562464b80260f4e2b6e403416b2315, but no
spec was added for it.