Add support for old versions of Capybara::RackTest::Form::NilUploadedFile
Capybara subclasses Rack::Test::UploadedFile, overrides initialize
and redefines a bunch of methods.  This is a terrible approach,
but as one of the goals is to keep capybara working with rack-test,
this hack is needed for compatibility.

Note that Capybara master already defines set_encoding and append_to,
so this is only needed for Capybara 3.37.1 and older.

Fixes #301