gemspec, spec_helper: drop unnecessary require_relative
The first line of `spec/spec_helper.rb` already adds the local lib/
directory to the Ruby load path, so this explicit require relative is
not really needed. For rack-test.gemspec, add lib/ to the load path.

Dropping the require_relative statements pointing into the local lib/
directory allows us to run the test suite against the files installed
system-wide, and that's an important QA tool for Debian.